test(structure): isolate transaction compatibility guards
