+++ docs/architecture/CODEBASE-REVIEW-LEDGER.md
@@ -10,6 +10,7 @@ Companion process document:
 
 | Review ID | Date | Scope / Pattern | Status | Findings | Actions Taken | Follow-up | Evidence / Sign-off |
 |---|---|---|---|---|---|---|---|
+| CR-1597 | 2026-07-15 | Transaction compatibility test ownership | Fixed locally; broader #719 campaign remains open | The cross-layer legacy calculator import and retired event-mapper confinement guard remained at the service-test root with repeated path setup. | Moved and renamed it under the transaction `architecture` test package, consolidated root constants, added owner/retired-path assertions, and updated repository context. | Organize image packaging and HTTP health tests separately; do not mix them into architecture governance. | Evidence: [CR-1597-TRANSACTION-COMPATIBILITY-TEST-OWNERSHIP.md](./codebase-reviews/CR-1597-TRANSACTION-COMPATIBILITY-TEST-OWNERSHIP.md); three focused and 835 package tests plus full architecture, docs, Ruff, and diff evidence recorded before commit. |
 | CR-1596 | 2026-07-15 | Transaction domain structure test ownership | Fixed locally; broader #719 campaign remains open | Cost-basis and processing-stage structure guards remained at the service-test root, and the cost-basis standard pinned the stale path. | Moved each guard beside its matching domain package, corrected repository-root resolution, added target/retired-path assertions, and reconciled standard/context truth. | Organize compatibility, image, and web tests separately; do not create a mixed governance package. | Evidence: [CR-1596-TRANSACTION-DOMAIN-STRUCTURE-TEST-OWNERSHIP.md](./codebase-reviews/CR-1596-TRANSACTION-DOMAIN-STRUCTURE-TEST-OWNERSHIP.md); 11 focused and 834 package tests plus domain-layer, docs, Ruff, and diff evidence recorded before commit. |
 | CR-1595 | 2026-07-15 | Transaction application test ownership | Fixed locally; broader #719 campaign remains open | Process-transaction and booked-transaction replay use-case tests remained at the service-test root after production application ownership was established; concurrency and risk coverage evidence pinned the old process-test path. | Moved both behavior tests under the application boundary with production-aligned names, added docstrings and no-return guards, and reconciled concurrency/risk evidence plus repository context. | Organize domain-structure, compatibility, image, and web tests separately; do not create a mixed governance package. | Evidence: [CR-1595-TRANSACTION-APPLICATION-TEST-OWNERSHIP.md](./codebase-reviews/CR-1595-TRANSACTION-APPLICATION-TEST-OWNERSHIP.md); 24 focused and 832 package tests plus concurrency, risk, architecture, docs, Ruff, and diff evidence recorded before commit. |
 | CR-1594 | 2026-07-15 | Import-safe changed coverage discovery | Fixed locally; broader #749 campaign remains open | Exact dotted-module coverage targets caused `pytest-cov` discovery to execute application package initializers before collection, producing duplicate Prometheus collector registration; removing that side effect exposed a real transaction-readiness repository coverage gap. | Changed governed coverage collection to deduplicated nearest-source directories while retaining exact-file JSON enforcement; added complete transaction-readiness stage repository behavior tests. | Keep #749 open for broader change-aware lane selection; do not weaken coverage thresholds or exact changed-file evidence. | Evidence: [CR-1594-IMPORT-SAFE-CHANGED-COVERAGE-DISCOVERY.md](./codebase-reviews/CR-1594-IMPORT-SAFE-CHANGED-COVERAGE-DISCOVERY.md); focused coverage-helper/readiness tests, 100% adapter line/branch proof, Ruff, and complete combined-gate evidence recorded before commit. |
